Fixed statue bug (I think?)
I believe the statue bug was caused by the nextthink value in walkmonster_start being set to 0 because of the random() function returning 0. That would explain why the bug is so rare. But because the bug is so rare, I can't easily reproduce it for testing. So hopefully it's fixed?