Made some changes to how deathmatch level changing
Deathmatch mode will automatically switch to the next level after the round ends. It will now go to the end map after the four episodes. It also picks randomly between exits if the map has two exits. And if a map doesn't have a trigger_changelevel entity, the map repeats instead of going to the start map.