Map changes and new texture
Removed second encounter with boss, lighting gun and the enemies in that area, and in its place added a new area as a break.

Added a new leaves texture that is used in this new area.

Reduced dificulty of the fight in general, with less enemies and added extra armor and health on normal and easy.