MAPS: e2m4 death knight fix
Sort of fixes the problem, in fteqw, where the death knight on the big eye arena wouldnt jump off the func_train. Now it has a wait value in that path_corner so the death knight can now jump off. This may lead to some inconsistensies to when he jumps off but its fine enough